+++ src/bin/pass-it-on-server.rs
@@ -1,3 +1,4 @@
+use clap::Parser;
 use log::{error, info, LevelFilter};
 use pass_it_on::start_server;
 use pass_it_on::Error;
@@ -6,9 +7,21 @@ use std::path::PathBuf;
 
 const LOG_TARGET: &str = "pass_it_on_server";
 
+#[derive(Parser, Debug)]
+#[clap(name = "pass-it-on-server", author, version, about = "Pass-it-on server binary", long_about = None)]
+struct CliArgs {
+    /// Path to pass-it-on server configuration file
+    #[clap(short, long, value_parser)]
+    configuration: Option<PathBuf>,
+    /// Set the logging level [default: Info]
+    #[clap(short, long, value_parser)]
+    log_level: Option<LevelFilter>,
+}
+
 #[tokio::main]
 async fn main() {
-    let module_log_level = module_debug_level();
+    let cli = CliArgs::parse();
+    let module_log_level = cli.log_level.unwrap_or(LevelFilter::Info);
     simple_logger::SimpleLogger::new()
         .with_level(LevelFilter::Off)
         .env()
@@ -18,19 +31,18 @@ async fn main() {
         .init()
         .unwrap();
 
-    if let Err(error) = run().await {
+    if let Err(error) = run(cli).await {
         error!(target: LOG_TARGET, "{}", error)
     }
 }
-async fn run() -> Result<(), Error> {
+async fn run(args: CliArgs) -> Result<(), Error> {
     // Setup default directories
-    let default_config_path = directories::ProjectDirs::from("net", "pass-on", "pass-on").unwrap();
+    let default_config_path = directories::ProjectDirs::from("com", "pass-it-on", "pass-it-on-server").unwrap();
 
     // Parse Config file
     let server_config = {
-        let args: Vec<String> = std::env::args().collect();
-        let config_path = match args.get(1) {
-            Some(path) => PathBuf::from(path),
+        let config_path = match args.configuration {
+            Some(path) => path,
             None => PathBuf::from(default_config_path.config_dir()).join("server.toml"),
         };
 
@@ -39,23 +51,4 @@ async fn run() -> Result<(), Error> {
     };
 
     start_server(server_config, None).await
-}
-
-fn module_debug_level() -> LevelFilter {
-    let args: Vec<String> = std::env::args().collect();
-    let limit = args.len() - 1;
-
-    for i in 0..limit {
-        if args.get(i).unwrap().contains("--log-level") && i < limit {
-            match args.get(i + 1).unwrap().trim().to_ascii_lowercase().as_str() {
-                "trace" => return LevelFilter::Trace,
-                "debug" => return LevelFilter::Debug,
-                "info" => return LevelFilter::Info,
-                "warn" => return LevelFilter::Warn,
-                "error" => return LevelFilter::Error,
-                _ => (),
-            }
-        }
-    }
-    LevelFilter::Info
 }