+++ src/interfaces/http/http_server.rs
@@ -3,26 +3,45 @@ use crate::LIB_LOG_TARGET;
 use log::{info, warn};
 use std::convert::Infallible;
 use std::net::SocketAddr;
+use std::path::Path;
 use tokio::sync::{mpsc, watch};
 use warp::http::StatusCode;
 use warp::Filter;
 
 const DEFAULT_BODY_LIMIT: u64 = 1024 * 1024;
 
-pub(super) async fn start_monitoring(tx: mpsc::Sender<String>, shutdown: watch::Receiver<bool>, socket: SocketAddr) {
+pub(super) async fn start_monitoring<P: AsRef<Path>>(
+    tx: mpsc::Sender<String>,
+    shutdown: watch::Receiver<bool>,
+    socket: SocketAddr,
+    tls: bool,
+    tls_cert_path: Option<P>,
+    tls_key_path: Option<P>,
+) {
     let mut shutdown_rx = shutdown.clone();
     let sender = warp::any().map(move || tx.clone());
 
     let filter2 = warp::path!("notification").and(notification_json_body()).and(sender).and_then(receive_notification);
 
-    info!(target: LIB_LOG_TARGET, "Setting up Interface: HttpSocket on -> {}", socket);
-    //warp::serve(filter2).run(socket).await;
-    //let x = shutdown_rx.changed().await.ok();
-    let (_address, server) = warp::serve(filter2).bind_with_graceful_shutdown(socket, async move {
-        shutdown_rx.changed().await.ok().unwrap_or_default();
-    });
-
-    server.await;
+    info!(target: LIB_LOG_TARGET, "Setting up Interface: HttpSocket on -> {} | TLS Enabled -> {}", socket, tls);
+    match tls {
+        true => {
+            let (_address, server) = warp::serve(filter2)
+                .tls()
+                .cert_path(tls_cert_path.unwrap().as_ref())
+                .key_path(tls_key_path.unwrap().as_ref())
+                .bind_with_graceful_shutdown(socket, async move {
+                    shutdown_rx.changed().await.ok().unwrap_or_default();
+                });
+            server.await;
+        }
+        false => {
+            let (_address, server) = warp::serve(filter2).bind_with_graceful_shutdown(socket, async move {
+                shutdown_rx.changed().await.ok().unwrap_or_default();
+            });
+            server.await;
+        }
+    };
 }
 
 async fn receive_notification(